+++ src/components/usuarios/VerUsuarios.js
@@ -3,13 +3,15 @@ import TablaGestion from '../tablas/TablaGestion';
 import UsuarioContext from '../../context/usuarios/usuarioContext';
 import AlertaContext from '../../context/alertas/alertaContext';
 import { columnasUsuarios } from '../util/Columnas';
-
+import AuthContext from '../../context/autenticacion/authContext';
 const VerUsuarios = () => {
     
     const usuarioContext = useContext(UsuarioContext); 
     const {obtenerUsuarios, usuarios ,eliminarUsuario,editarUsuario} = usuarioContext;
     const alertaContext = useContext(AlertaContext);
-    const {alerta} = alertaContext;
+    const {alerta,mostrarAlerta} = alertaContext;
+    const authContext = useContext(AuthContext);
+    const {usuario} = authContext;
     
     useEffect(() => {
         obtenerUsuarios();
@@ -18,18 +20,28 @@ const VerUsuarios = () => {
 
 
     const actualizarFila = (newData, oldData, resolve, reject) => {
-        console.log('paso por aca');
-        editarUsuario(newData);
-        setTimeout(() => {
-            resolve()
-        }, 3000);
+        if(usuario && (usuario.cargo == 'jefeproduccion' || usuario.cargo == 'admin') ){
+            editarUsuario(newData);
+            setTimeout(() => {
+                resolve()
+            }, 3000);
+
+        }else{
+            mostrarAlerta('No tienes los permisos para realizar la accion', 'alerta-error');
+            reject();
+        }
         
     }
     const eliminarFila = async (oldData, resolve, reject) => {
-        eliminarUsuario(oldData);
-        setTimeout(() => {
-            resolve()
-        }, 3000);
+        if(usuario && (usuario.cargo == 'jefeproduccion' || usuario.cargo == 'admin') ){
+            eliminarUsuario(oldData);
+            setTimeout(() => {
+                resolve()
+            }, 3000);
+        }else{
+            mostrarAlerta('No tienes los permisos para realizar la accion', 'alerta-error');
+            reject();
+        }
     }
     if(usuarios.length === 0){
         return null;
